Rewrite `get_fn_id_for_return_block`
